# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CIAM Microservice API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Mandai Wildlife Group"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AppId": {
        "type": "apiKey",
        "in": "header",
        "name": "mwg-app-id",
        "description": "Application ID for client identification. Required for all API calls."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authorized clients. Required when api_key_validation switch is enabled."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from login or passwordless authentication."
      }
    },
    "schemas": {
...
      "PasswordlessSendRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address to send OTP",
            "example": "user@example.com"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        }
      },
      "PasswordlessSendResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "failed"],
            "example": "success"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 200
              },
              "message": {
                "type": "string",
                "example": "OTP sent successfully"
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              }
            }
          }
        }
      },
      "PasswordlessVerifyRequest": {
        "type": "object",
        "required": ["email", "otp"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "otp": {
            "type": "string",
            "description": "One-time password received via email",
            "example": "123456"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        }
      },
      "PasswordlessVerifyResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "failed"],
            "example": "success"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 200
              },
              "mwgCode": {
                "type": "string",
                "example": "MWG_CIAM_OTP_VERIFY_SUCCESS"
              },
              "message": {
                "type": "string",
                "example": "OTP verified successfully"
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "mandaiId": {
                "type": "string",
                "example": "MWG-12345678"
              },
              "accessToken": {
                "type": "string",
                "description": "JWT access token",
                "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6..."
              },
              "refreshToken": {
                "type": "string",
                "description": "Refresh token",
                "example": "eyJjdHkiOiJKV1QiLCJlbmMiOi..."
              },
              "idToken": {
                "type": "string",
                "description": "ID token",
                "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6..."
              },
              "expiresIn": {
                "type": "integer",
                "description": "Token expiry time in seconds",
                "example": 3600
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["failed"],
            "example": "failed"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 400
              },
              "mwgCode": {
                "type": "string",
                "example": "MWG_CIAM_PARAMS_ERR"
              },
              "message": {
                "type": "string",
                "example": "Wrong parameters"
              },
              "error": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["failed"],
            "example": "failed"
          },
          "statusCode": {
            "type": "integer",
            "example": 401
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 401
              },
              "mwgCode": {
                "type": "string",
                "example": "MWG_CIAM_UNAUTHORIZED"
              },
              "message": {
                "type": "string",
                "example": "Unauthorized"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["failed"],
            "example": "failed"
          },
          "statusCode": {
            "type": "integer",
            "example": 500
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 500
              },
              "mwgCode": {
                "type": "string",
                "example": "MWG_CIAM_INTERNAL_SERVER_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Internal Server Error"
              }
            }
          }
        }
      },
...
      "LanguageCode": {
        "type": "string",
        "enum": ["en", "ja", "kr", "zh"],
        "description": "Language code for localized messages",
        "example": "en"
      },
...
      "MembershipCheckRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address to check membership",
            "example": "user@example.com"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        }
      },
      "MembershipCheckResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "failed"],
            "example": "success"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 200
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "mandaiId": {
                "type": "string",
                "example": "MWG-12345678"
              },
              "hasMembership": {
                "type": "boolean",
                "example": true
              },
              "membershipDetails": {
                "type": "object",
                "properties": {
                  "categoryType": {
                    "type": "string",
                    "example": "Annual Pass"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  },
                  "status": {
                    "type": "string",
                    "example": "active"
                  }
                }
              }
            }
          }
        }
      },
...
      "MyAccountMembershipRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "mandaiId": {
            "type": "string",
            "description": "User's Mandai ID",
            "example": "MWG-12345678"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        }
      },
      "MyAccountMembershipResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "failed"],
            "example": "success"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 200
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "mandaiId": {
                "type": "string",
                "example": "MWG-12345678"
              },
              "firstName": {
                "type": "string",
                "example": "John"
              },
              "lastName": {
                "type": "string",
                "example": "Doe"
              },
              "membershipDetails": {
                "type": "object",
                "properties": {
                  "categoryType": {
                    "type": "string",
                    "example": "Annual Pass"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  }
                }
              }
            }
          }
        }
      },
      "DeleteAccountRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "mandaiId": {
            "type": "string",
            "description": "User's Mandai ID",
            "example": "MWG-12345678"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        }
      },
      "DeleteAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "failed"],
            "example": "success"
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "membership": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 200
              },
              "message": {
                "type": "string",
                "example": "Account deleted successfully"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "AppId": []
    }
  ],
  "paths": {
...
    "/v1/ciam/users/memberships": {
      "post": {
        "summary": "Check membership by email",
        "description": "Check if a user has an active membership by their email address",
        "tags": ["Memberships"],
        "security": [
          {
            "AppId": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCheckRequest"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Membership check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipCheckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/ciam/auth/passwordless/send": {
      "post": {
        "summary": "Send OTP email",
        "description": "Send a one-time password to the user's email for passwordless authentication",
        "tags": ["Auth"],
        "security": [
          {
            "AppId": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordlessSendRequest"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordlessSendResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - rate limit exceeded or email invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/ciam/auth/passwordless/session": {
      "post": {
        "summary": "Verify OTP",
        "description": "Verify the one-time password and create a session. Returns JWT tokens on success.",
        "tags": ["Auth"],
        "security": [
          {
            "AppId": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordlessVerifyRequest"
              },
              "example": {
                "email": "user@example.com",
                "otp": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully, session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordlessVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/ciam/auth/membership": {
      "post": {
        "summary": "Retrieve membership",
        "description": "Retrieve membership details for the authenticated user",
        "tags": ["Auth"],
        "security": [
          {
            "AppId": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyAccountMembershipRequest"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Membership details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyAccountMembershipResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user account",
        "description": "Delete the authenticated user's account and all associated data",
        "tags": ["Auth"],
        "security": [
          {
            "AppId": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    }
  }
}